fix: resolve compilation errors in timeline editor
- horizontalScroll is a Modifier extension, not a layout function:
  wrap TimelineView in Box(Modifier.horizontalScroll(scrollState))
- Double.toDp() doesn't exist: add .toFloat() before .toDp()
  in timelineWidthPx calculation
- Add missing Box import

+++ app/src/main/java/com/t2v/ui/screens/audioeditor/TimelineView.kt
@@ -61,7 +61,7 @@ fun TimelineView(
     val totalDurationMs = listOf(voiceClips, musicClips, soundClips).flatten()
         .maxOfOrNull { it.timelineStartMs + clipDurationMs(it) } ?: 0L
     val totalDurationSec = (totalDurationMs / 1000.0).coerceAtLeast(5.0)
-    val timelineWidthPx = with(density) { (totalDurationSec * pixelsPerSecond).toDp() }
+    val timelineWidthPx = with(density) { (totalDurationSec * pixelsPerSecond).toFloat().toDp() }
     val rulerHeight = 28.dp
     val laneHeight = 56.dp
     val laneGap = 4.dp
